+++ src/vue/forms/TransferForm.vue
@@ -18,11 +18,12 @@
 
       <template
         v-else-if="view.mode === VIEW_MODES.submit ||
-          view.mode === VIEW_MODES.confirm">
+          view.mode === VIEW_MODES.confirm"
+      >
         <form
+          id="transfer-form"
           @submit.prevent="processTransfer"
-          v-if="view.mode === VIEW_MODES.submit ||
-            view.mode === VIEW_MODES.confirm">
+        >
           <div class="app__form-row">
             <div class="app__form-field">
               <select-field
@@ -92,137 +93,15 @@
             </div>
           </div>
 
-          <transition name="app__fade-in">
-            <div
-              class="transfer__fee-box"
-              v-if="isFeesLoaded">
-              <h3 class="transfer__fee-box-heading">
-                {{ 'transfer-form.sender-fees' | globalize }}
-              </h3>
-
-              <!-- eslint-disable-next-line -->
-              <template v-if=" +fees.source.fixed || +fees.source.calculatedPercent || form.isPaidForRecipient ">
-                <p
-                  class="transfer__fee"
-                  v-if="fees.source.fixed">
-                  - {{ fees.source.fixed | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.sender-fixed-fee' | globalize }}
-                  </span>
-                </p>
-
-                <p
-                  class="transfer__fee"
-                  v-if="fees.source.calculatedPercent">
-                  - {{ fees.source.calculatedPercent | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.sender-percent-fee' | globalize }}
-                  </span>
-                </p>
-
-                <p
-                  class="transfer__fee"
-                  v-if="form.isPaidForRecipient && +fees.destination.fixed">
-                  - {{ fees.destination.fixed | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.recipient-fixed-fee' | globalize }}
-                  </span>
-                </p>
-
-                <!-- eslint-disable max-len -->
-                <p
-                  class="transfer__fee"
-                  v-if="form.isPaidForRecipient && +fees.destination.calculatedPercent">
-                  - {{ fees.destination.calculatedPercent | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.recipient-percent-fee' | globalize }}
-                  </span>
-                </p>
-                <!-- eslint-enable max-len -->
-              </template>
-
-              <template v-else>
-                <p class="transfer__no-fee-msg">
-                  {{ 'transfer-form.source-no-fees' | globalize }}
-                </p>
-              </template>
-
-              <h3 class="transfer__fee-box-heading">
-                {{ 'transfer-form.recipient-fees' | globalize }}
-              </h3>
-
-              <!-- eslint-disable-next-line max-len -->
-              <template v-if="(+fees.destination.fixed || +fees.destination.calculatedPercent) && !form.isPaidForRecipient">
-                <p
-                  class="transfer__fee"
-                  v-if="fees.destination.fixed">
-                  - {{ fees.destination.fixed | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.recipient-fixed-fee' | globalize }}
-                  </span>
-                </p>
-
-                <p
-                  class="transfer__fee"
-                  v-if="fees.destination.calculatedPercent">
-                  - {{ fees.destination.calculatedPercent | formatNumber }}
-                  {{ form.asset.code }}
-                  <span class="transfer__fee-type">
-                    {{ 'transfer-form.recipient-percent-fee' | globalize }}
-                  </span>
-                </p>
-              </template>
-
-              <template v-else>
-                <p class="transfer__no-fee-msg">
-                  {{ 'transfer-form.recipient-no-fees' | globalize }}
-                </p>
-              </template>
-
-              <h3 class="transfer__fee-box-heading">
-                {{ 'transfer-form.total' | globalize }}
-              </h3>
-
-              <p class="transfer__fee">
-                - {{ totalSenderFee | formatNumber }}
-                {{ form.asset.code }}
-                <span class="transfer__fee-type">
-                  {{ 'transfer-form.total-sender-fee' | globalize }}
-                </span>
-              </p>
-
-              <p class="transfer__fee">
-                - {{ totalReceiverFee | formatNumber }}
-                {{ form.asset.code }}
-                <span class="transfer__fee-type">
-                  {{ 'transfer-form.total-receiver-fee' | globalize }}
-                </span>
-              </p>
-
-              <p class="transfer__fee">
-                - {{ totalAmount | formatNumber }} {{ form.asset.code }}
-                <span class="transfer__fee-type">
-                  {{ 'transfer-form.total-amount' | globalize }}
-                </span>
-              </p>
-
-              <!-- eslint-disable max-len -->
-              <div
-                class="app__form-row"
-                v-if="+fees.destination.fixed || +fees.destination.calculatedPercent"
-              >
-                <tick-field v-model="form.isPaidForRecipient">
-                  {{ 'transfer-form.pay-fees-for-recipient' | globalize }}
-                </tick-field>
-              </div>
-              <!-- eslint-enable max-len -->
-            </div>
-          </transition>
+          <div
+            class="transfer__fee-box"
+            v-if="isFeesLoaded && view.mode === VIEW_MODES.confirm"
+          >
+            <fees-renderer
+              :fees-collection="fees"
+              :paid-for-destination.sync="form.isPaidForRecipient"
+            />
+          </div>
 
           <div class="app__form-actions">
             <button
@@ -231,6 +110,7 @@
               type="submit"
               class="app__form-submit-btn app__button-raised"
               :disabled="formMixin.isDisabled"
+              form="transfer-form"
             >
               {{ 'transfer-form.continue-btn' | globalize }}
             </button>
@@ -265,16 +145,11 @@ import Loader from '@/vue/common/Loader'
 import FormMixin from '@/vue/mixins/form.mixin'
 import IdentityGetterMixin from '@/vue/mixins/identity-getter'
 import { vueRoutes } from '@/vue-router/routes'
-
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { mapGetters, mapActions } from 'vuex'
 import { vuexTypes } from '@/vuex'
-import {
-  base,
-  PAYMENT_FEE_SUBTYPES,
-  FEE_TYPES,
-} from '@tokend/js-sdk'
-import { MathUtil } from '@/js/utils'
+import { base, FEE_TYPES } from '@tokend/js-sdk'
+import FeesMixin from '@/vue/common/fees/fees.mixin'
 import config from '@/config'
 import { api } from '@/api'
 import { Bus } from '@/js/helpers/event-bus'
@@ -300,7 +175,11 @@ export default {
   components: {
     Loader,
   },
-  mixins: [FormMixin, IdentityGetterMixin],
+  mixins: [
+    FormMixin,
+    IdentityGetterMixin,
+    FeesMixin,
+  ],
   props: {
     assetToTransfer: { type: String, default: '' },
   },
@@ -312,19 +191,10 @@ export default {
       subject: '',
       isPaidForRecipient: false,
     },
+    fees: {},
     view: {
       mode: VIEW_MODES.submit,
       opts: {},
-    },
-    fees: {
-      source: {
-        fixed: '',
-        calculatedPercent: '',
-      },
-      destination: {
-        fixed: '',
-        calculatedPercent: '',
-      },
     },
     isLoaded: false,
     isLoadingFailed: false,
@@ -332,6 +202,7 @@ export default {
     VIEW_MODES,
     vueRoutes,
     config,
+    FEE_TYPES,
   }),
   validations () {
     return {
@@ -360,25 +231,6 @@ export default {
     balance () {
       return this.accountBalances
         .find(i => i.asset.code === this.form.asset.code) || {}
-    },
-    totalSenderFee () {
-      return MathUtil.add(
-        this.fees.source.fixed,
-        this.fees.source.calculatedPercent
-      )
-    },
-    totalReceiverFee () {
-      return MathUtil.add(
-        this.fees.destination.fixed,
-        this.fees.destination.calculatedPercent
-      )
-    },
-    totalAmount () {
-      const fees = this.form.isPaidForRecipient
-        ? MathUtil.add(this.totalSenderFee, this.totalReceiverFee)
-        : this.totalSenderFee
-
-      return MathUtil.add(fees, this.form.amount)
     },
   },
   async created () {
@@ -418,22 +270,27 @@ export default {
       try {
         const recipientAccountId =
           await this.getCounterparty(this.form.recipient)
-        const fees = await this.getFees(recipientAccountId)
-        this.fees = fees
+
+        this.fees = await this.calculateFees({
+          assetCode: this.form.asset.code,
+          amount: this.form.amount,
+          recipientAccountId: recipientAccountId,
+          senderAccountId: this.accountId,
+          type: FEE_TYPES.paymentFee,
+        })
         this.isFeesLoaded = true
 
         const opts = {
           amount: this.form.amount,
           destinationAccountId: recipientAccountId,
-          destinationFixedFee: fees.destination.fixed,
-          destinationPercentFee: fees.destination.calculatedPercent,
-          destinationFeeAsset: this.form.asset.code,
+          destinationFixedFee: this.fees.destinationFee.fixed,
+          destinationPercentFee: this.fees.destinationFee.calculatedPercent,
+          destinationFeeAsset: this.form.asset,
           sourceBalanceId: this.balance.id,
-          sourceFixedFee: fees.source.fixed,
-          sourcePercentFee: fees.source.calculatedPercent,
-          sourceFeeAsset: this.form.asset.code,
+          sourceFixedFee: this.fees.sourceFee.fixed,
+          sourcePercentFee: this.fees.sourceFee.calculatedPercent,
+          sourceFeeAsset: this.form.asset,
           subject: this.form.subject,
-          assetCode: this.form.asset.code,
         }
         this.updateView(VIEW_MODES.confirm, opts)
       } catch (error) {
@@ -446,45 +303,6 @@ export default {
         return this.getAccountIdByEmail(recipient)
       } else {
         return recipient
-      }
-    },
-    async getFees (recipientAccountId) {
-      const [senderFees, recipientFees] = await Promise.all([
-        this.loadPaymentFee({
-          asset: this.form.asset.code,
-          amount: this.form.amount,
-          account: this.accountId,
-          subtype: PAYMENT_FEE_SUBTYPES.outgoing,
-        }),
-        this.loadPaymentFee({
-          asset: this.form.asset.code,
-          amount: this.form.amount,
-          account: recipientAccountId,
-          subtype: PAYMENT_FEE_SUBTYPES.incoming,
-        }),
-      ])
-      return {
-        source: senderFees,
-        destination: recipientFees,
-      }
-    },
-    async loadPaymentFee ({ asset, amount, account, subtype }) {
-      try {
-        const baseEndpoint = `/v3/accounts/${account}/calculated_fees`
-        const params = [
-          `asset=${asset}`,
-          `fee_type=${FEE_TYPES.paymentFee}`,
-          `subtype=${subtype}`,
-          `amount=${amount}`,
-        ]
-
-        const endpoint = `${baseEndpoint}?${params.join('&')}`
-        const { data: fees } = await api.get(endpoint)
-
-        return fees
-      } catch (e) {
-        ErrorHandler.process(e)
-        return {}
       }
     },
     buildPaymentOperation () {
@@ -535,9 +353,7 @@ export default {
 @import '~@scss/variables';
 
 .transfer__fee-box {
-  margin-top: 4rem;
   padding-top: 4rem;
-  border-top: 0.1rem dashed $col-text-field-hint-inactive;
 }
 
 .transfer__fee-box-heading {