+++ src/vue/forms/market-orders/YourTradeOfferForm.vue
@@ -60,25 +60,36 @@
 
     <div class="app__form-row">
       <div class="app__form-field">
-        <readonly-field
-          :label="
-            'your-trade-offer-form.total-amount-lbl' | globalize({
-              asset: assetPair.quote
-            })
-          "
-          :value="{
-            value: quoteAmount,
-            currency: assetPair.quote,
-          } | formatMoney"
-          :error-message="getFieldErrorMessage(
-            'quoteAmount',
-            {
-              available: quoteAssetBalance,
-              from: config.MIN_AMOUNT,
-              to: config.MAX_AMOUNT,
-            }
-          )"
-        />
+        <template v-if="isFeesLoaded">
+          <readonly-field
+            :label="
+              'your-trade-offer-form.total-amount-lbl' | globalize({
+                asset: assetPair.quote
+              })
+            "
+            :value="{
+              value: quoteAmount,
+              currency: assetPair.quote,
+            } | formatMoney"
+            :error-message="getFieldErrorMessage(
+              'quoteAmount',
+              {
+                available: quoteAssetBalance,
+                from: config.MIN_AMOUNT,
+                to: config.MAX_AMOUNT,
+              }
+            )"
+          />
+
+          <fees-renderer
+            class="your-trade-offer-form__fees"
+            :fees-collection="fees"
+          />
+        </template>
+
+        <template v-else>
+          <loader message-id="your-trade-offer-form.loading-msg" />
+        </template>
       </div>
     </div>
 
@@ -98,7 +109,7 @@
           type="submit"
           @click="setUpdateSubmitMode"
           class="app__button-raised your-trade-offer-form__btn"
-          :disabled="formMixin.isDisabled"
+          :disabled="formMixin.isDisabled || !isFeesLoaded"
         >
           {{ 'your-trade-offer-form.update-order-btn' | globalize }}
         </button>
@@ -115,13 +126,22 @@
       </div>
     </template>
   </form>
+
+  <loader v-else message-id="your-trade-offer-form.loading-msg" />
 </template>
 
 <script>
+import debounce from 'lodash/debounce'
+
 import ReadonlyField from '@/vue/fields/ReadonlyField'
+import FeesRenderer from '@/vue/common/fees/FeesRenderer'
+import Loader from '@/vue/common/Loader'
 
 import FormMixin from '@/vue/mixins/form.mixin'
 import OfferManagerMixin from '@/vue/mixins/offer-manager.mixin'
+import FeesMixin from '@/vue/common/fees/fees.mixin'
+
+import { FEE_TYPES } from '@tokend/js-sdk'
 
 import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
@@ -146,12 +166,19 @@ const SUBMIT_MODES = {
   update: 'update',
 }
 
+const FEES_LOADING_DELAY_MS = 300
+
 export default {
   name: 'your-trade-offer-form',
-  components: { ReadonlyField },
+  components: {
+    ReadonlyField,
+    FeesRenderer,
+    Loader,
+  },
   mixins: [
     FormMixin,
     OfferManagerMixin,
+    FeesMixin,
   ],
 
   props: {
@@ -165,6 +192,10 @@ export default {
       price: '',
       baseAmount: '',
     },
+    fees: {},
+    feesDebouncedRequest: null,
+    isFeesLoaded: false,
+    isFeesLoadFailed: false,
     isLoaded: false,
     submitMode: '',
     isFormSubmitting: false,
@@ -253,6 +284,7 @@ export default {
         quoteAmount: this.quoteAmount,
         price: this.form.price,
         isBuy: this.isBuy,
+        fee: this.fees.totalFee,
       }
     },
 
@@ -266,10 +298,21 @@ export default {
     },
   },
 
+  watch: {
+    'form.baseAmount' () {
+      this.tryLoadFees()
+    },
+
+    'form.price' () {
+      this.tryLoadFees()
+    },
+  },
+
   async created () {
     try {
       await this.loadBalances()
       this.populateForm()
+      await this.loadFees()
       this.isLoaded = true
     } catch (e) {
       ErrorHandler.processWithoutFeedback(e)
@@ -280,6 +323,35 @@ export default {
     populateForm () {
       this.form.baseAmount = this.offer.baseAmount
       this.form.price = this.offer.price
+    },
+
+    tryLoadFees () {
+      this.isFeesLoaded = false
+      this.isFeesLoadFailed = false
+
+      if (!this.feesDebouncedRequest) {
+        this.feesDebouncedRequest = debounce(
+          () => this.loadFees(),
+          FEES_LOADING_DELAY_MS
+        )
+      }
+      return this.feesDebouncedRequest()
+    },
+
+    async loadFees () {
+      try {
+        this.fees = await this.calculateFees({
+          assetCode: this.assetPair.quote,
+          amount: this.quoteAmount || 0,
+          senderAccountId: this.accountId,
+          type: FEE_TYPES.offerFee,
+        })
+
+        this.isFeesLoaded = true
+      } catch (e) {
+        this.isFeesLoadFailed = true
+        ErrorHandler.processWithoutFeedback(e)
+      }
     },
 
     setUpdateSubmitMode () {
@@ -327,5 +399,9 @@ export default {
 .your-trade-offer-form__btn {
   max-width: 16rem;
   width: 100%;
+}
+
+.your-trade-offer-form__fees {
+  margin-top: 1rem;
 }
 </style>