+++ src/vue/modules/create-sale-form/mixins/manage-sale-request.mixin.js
@@ -1,10 +1,10 @@
-import UploadDocumentsMixin from './upload-documents.mixin'
 import ManageSaleDescriptionMixin from './manage-sale-description.mixin'
 
 import { base, SALE_TYPES } from '@tokend/js-sdk'
 
 import { api } from '@/api'
 
+import { uploadDocument } from '@/js/helpers/upload-documents'
 import { CreateSaleRequest } from '../wrappers/create-sale-request'
 import { DateUtil } from '@/js/utils'
 
@@ -19,7 +19,7 @@ const EMPTY_DOCUMENT = {
 }
 
 export default {
-  mixins: [UploadDocumentsMixin, ManageSaleDescriptionMixin],
+  mixins: [ManageSaleDescriptionMixin],
   data: _ => ({
     saleDescriptionBlobId: '',
   }),
@@ -72,10 +72,7 @@ export default {
     },
 
     async submitCreateSaleRequest (accountId) {
-      const saleDocuments = [
-        this.shortBlurbStepForm.saleLogo,
-      ]
-      await this.uploadDocuments(saleDocuments, accountId)
+      await uploadDocument(this.shortBlurbStepForm.saleLogo)
       this.saleDescriptionBlobId = await this.createSaleDescriptionBlob(
         this.fullDescriptionStepForm.description,
         accountId