+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for further information about branching and tagging conventions.
 
 ## [Unreleased]
 #### Added
+- Placeholder of empty list for Trade, Issuance requests and Requests pages
 - Tooltip directive
 - Assets vuex module for loading and storing all the assets
 - Whitelisted detail on sale related modules
@@ -16,6 +17,10 @@ for further information about branching and tagging conventions.
 - Usage of Unix line endings rule
 - Browser compatibiility check and unsupported browser page
 - Password toggle button for input fields with type="password"
+- Fees mixin, renderer & wrappers for unified displaying fees
+- Displaying fees on issuance form & trade forms
+- State icon to the status message
+- Rotation of status message close icon on hovering
 - Cap asset field to the create sale form
 
 #### Changed
@@ -28,6 +33,10 @@ for further information about branching and tagging conventions.
 - Now hiding "My Sales" tab for non-corporate accounts
 - Update "Learn more about pre-issuance" link style on sale creation form
 - Now including offer fees to account's balance on invest form
+- Now using `DocumentsManager` from SDK to upload documents to the storage
+- Now getting document URL using documentsManager
+- Now using light background for status messages with the type
+  "success" of "info"
 - Now using cap asset as default quote asset for create sale request
 - Now creating quote asset balances on create sale form if they don't exist
 - Now displaying accepted investment assets as base assets of pairs where